requirements.txt
flask
google-auth==2.11.0
google-auth-oauthlib==0.5.2
google-api-python-client==2.60.0
python-dotenv==0.19.2